+++ tests/data/expected/sql/224913000.sql
@@ -3,7 +3,7 @@ SELECT mmsi, lat, lon,
   shipname,
   speed
 
-FROM TABLE_DATE_RANGE([pipeline_classify_logistic_661b.], 
+FROM TABLE_DATE_RANGE([pipeline_740__classify.],
   TIMESTAMP('2015-11-19'), TIMESTAMP('2015-11-20'))
 
 where mmsi in (224913000)